added writing of trivial header file; added trivial test case

+++ program/main.cpp
@@ -3,5 +3,5 @@
 
 int main(int argc, char **argv)
 {
-  return f2h::program(argc, argv);
+  return static_cast<int>(f2h::program(argc, argv));
 }
